SOL-1802: Create internal amqpNetLite logging
Create a amqp NetLite TraceListener and set the amqp logging
to maximum (TRACE | VERBOSE ). This is not a large overhead because
amqpNetLite always evaluates all arguments to Trace statements, so we
are just adding a method call overhead.  In the NMS TraceListener call
the appropriate Tracer method.

Initialize Tracer in UnitTest and HelloWorld appropriately.