SOL-3669: Creating/Deleting Durable Topic Endpoint on VMR causes 'already bound'
The already bound exception is being thrown from the AMQP ReceiverLink constructor. The NMS MessageConsumer.close() method does not block until the broker confirms the 'detach' frame. A subsequent attempt to create a MessageConsumer for the same subscription (topic-endpoint) caused the AMQP receiverLink constructor to throw as AMQP tracks the subscription until the detach confirm is received.

Fixed the NMS API to block in ConsumerClose() for the configured timeout.